LayoutTests/imported/w3c/web-platform-tests/pointerevents/persistentDeviceId/pointer-event-has-persistentdeviceid-from-pointer-event-init.tentative-expected.txt


FAIL PointerEvent.persistentDeviceId via PointerEventInit assert_equals: persistentDeviceId is populated expected (number) 1001 but got (undefined) undefined
FAIL No persistentDeviceId in PointerEventInit assert_equals: expected (number) 0 but got (undefined) undefined
